+++ index.js
@@ -122,9 +122,9 @@ exports.get_dovecot_response = function (connection, domain, email, cb) {
     }
   }
 
-  const socket_address = options.path
-    ? options.path
-    : `${options.host}:${options.port}`
+  //const socket_address = options.path
+  //  ? options.path
+  //  : `${options.host}:${options.port}`
   connection.transaction.results.add(plugin, {
     msg: `sock: ${options.host}:${options.port}`,
   })